Relocate process tree classes
