--- src/aiidalab_qe/app/result/components/status/process_tree/__init__.py
@@ -1,7 +1,0 @@
-from .model import SimplifiedProcessTreeModel
-from .tree import SimplifiedProcessTree
-
-__all__ = [
-    "SimplifiedProcessTree",
-    "SimplifiedProcessTreeModel",
-]
