--- src/aiidalab_qe/app/result/components/status/process_tree/model.py
@@ -1,8 +1,0 @@
-import traitlets as tl
-
-from aiidalab_qe.common.mixins import HasProcess
-from aiidalab_qe.common.mvc import Model
-
-
-class SimplifiedProcessTreeModel(Model, HasProcess):
-    clicked = tl.Unicode(None, allow_none=True)
